+++ shell/shopt/expand_aliases.sh
@@ -1,0 +1,13 @@
+#!/bin/bash
+set -e
+
+my_uname()
+{
+	echo "rtoax"
+}
+
+uname
+
+shopt -s expand_aliases
+alias uname=my_uname
+uname